+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix `attributeMapper` calls to pass the node as the first argument, not the attributes object.
+
 ## 0.0.4 (2026-08-23)
 
 - Update terminology to replace the term "content" with "node", to match the terminology used by rich text rendering libraries.